+++ routing/ff_router_trips/UnivFFviaFMTrips.cpp
@@ -846,7 +846,7 @@ void UnivFFviaFMTrips::drawLinesOnGrid(std::vector<Line> & wallArg, T * const ta
 
 template <typename T>
 void UnivFFviaFMTrips::drawLinesOnGrid(Line & line, T * const target, const T value)
-{   //no init, plz init elsewhere
+{ //no init, plz init elsewhere
     // i~x; j~y;
     //http://stackoverflow.com/questions/10060046/drawing-lines-with-bresenhams-line-algorithm
     //src in answer of "Avi"; adapted to fit this application
@@ -958,7 +958,7 @@ void UnivFFviaFMTrips::drawLinesOnWall(std::vector<Line> & wallArg, T * const ta
 
 template <typename T>
 void UnivFFviaFMTrips::drawLinesOnWall(Line & line, T * const target, const T value)
-{   //no init, plz init elsewhere
+{ //no init, plz init elsewhere
     // i~x; j~y;
     //http://stackoverflow.com/questions/10060046/drawing-lines-with-bresenhams-line-algorithm
     //src in answer of "Avi"; adapted to fit this application