+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 ## v0.8.5 [2019-07-12]
 
 ### Added
+- Added option for additional/optional output for speed, velocity(x,y), 
+final goal, intermediate goal, desired direction(x,y), spotlight, router, group. Thus,
+there is no need to run multiple simulation to obtain those values. Attention: File may
+become significantly larger!
 
 ### Changed